chore: remove legacy trade execution utility (#5225)
